Nachricht Controller/Repo/Service

+++ backend/spring/src/main/java/at/gradwohl/website/repository/nachricht/NachrichtRepository.java
@@ -3,5 +3,6 @@ package at.gradwohl.website.repository.nachricht;
 import at.gradwohl.website.model.nachricht.Nachricht;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NachrichtRepository extends JpaRepository<Nachricht, Integer> {
+public interface NachrichtRepository extends JpaRepository<Nachricht, Long> {
+    
 }
